added lights to platform io project
